add dogecoin price to bitcoin command

+++ bot.js
@@ -33,12 +33,12 @@ const processCommand = async (message) => {
   const fullCommand = message.content.substr(2);
   const splitCommand = fullCommand.split(' ');
   const primaryCommand = splitCommand[0];
-  const arguments = splitCommand.slice(1);
+  const cmdArguments = splitCommand.slice(1);
 
   switch (primaryCommand.toLowerCase()) {
     case 'bug':
-      if (arguments.length > 0) {
-        const bug = await arguments.join(' ');
+      if (cmdArguments.length > 0) {
+        const bug = await cmdArguments.join(' ');
         etrizDM(bug);
         message.channel.send('Thanks for finding this bug');
       } else {
@@ -46,19 +46,19 @@ const processCommand = async (message) => {
       }
       break;
     case 'help':
-      helpCommand(arguments, message);
+      helpCommand(cmdArguments, message);
       break;
     case 'next':
       const next = await db.getCommandData('next');
       message.channel.send(`Next meeting is ${next}`);
       break;
     case 'addcommand':
-      if (arguments.length === 0) {
+      if (cmdArguments.length === 0) {
         message.channel.send(`Try !!addcommand [command] [details]`);
       } else {
-        if (arguments.length > 1) {
-          const newCommand = arguments[0];
-          const temp = arguments.slice(1);
+        if (cmdArguments.length > 1) {
+          const newCommand = cmdArguments[0];
+          const temp = cmdArguments.slice(1);
           const details = await temp.join(' ');
           const addAttempt = await db.addNewCommand(newCommand, details);
           message.channel.send(`Added command !!${addAttempt}`);
@@ -68,11 +68,11 @@ const processCommand = async (message) => {
       }
       break;
     case 'set':
-      if (arguments.length === 0) {
+      if (cmdArguments.length === 0) {
         message.channel.send(`Try !!set [command] [details]`);
-      } else if (arguments[0].toLowerCase() === 'next') {
-        if (arguments.length > 1) {
-          let temp = arguments.slice(1);
+      } else if (cmdArguments[0].toLowerCase() === 'next') {
+        if (cmdArguments.length > 1) {
+          let temp = cmdArguments.slice(1);
           const next = temp.join(' ');
           const setAttempt = await db.setNext(next);
           // console.log(setAttempt);
@@ -80,10 +80,10 @@ const processCommand = async (message) => {
         } else {
           message.channel.send('Please supply details');
         }
-      } else if (arguments[0].toLowerCase() === 'activity') {
+      } else if (cmdArguments[0].toLowerCase() === 'activity') {
         if (message.author.id === etriz) {
-          client.user.setActivity(arguments.slice(2).join(' '), {
-            type: arguments[1].toUpperCase(),
+          client.user.setActivity(cmdArguments.slice(2).join(' '), {
+            type: cmdArguments[1].toUpperCase(),
           });
         } else {
           message.channel.send('This is a owner only command');
@@ -108,8 +108,8 @@ const processCommand = async (message) => {
   }
 };
 
-const helpCommand = (arguments, message) => {
-  arguments.length > 0
+const helpCommand = (cmdArguments, message) => {
+  cmdArguments.length > 0
     ? message.channel.send(`It looks like you need help with ${arguments}`)
     : message.channel.send('The commands are: !!next, !!set next, !!bitcoin, !!bug');
 };
@@ -117,14 +117,17 @@ const helpCommand = (arguments, message) => {
 const bitCoinCommand = async (message) => {
   const search = await message.channel.send('Searching ...');
   const res = await axios.get(
-    'https://api.coingecko.com/api/v3/simple/price?ids=ethereum%2Cbitcoin&vs_currencies=usd'
+    'https://api.coingecko.com/api/v3/simple/price?ids=ethereum%2Cbitcoin%2Cdogecoin&vs_currencies=usd'
   );
   const bitcoin = await res.data.bitcoin.usd;
   const eth = await res.data.ethereum.usd;
-  if (bitcoin && eth) {
+  const doge = await res.data.dogecoin.usd;
+  if (bitcoin && eth && doge) {
     search.delete();
     message.reply(
-      `Bitcoin price: ${formatCurrency(bitcoin)}\nEthereum price: ${formatCurrency(eth)}`
+      `Bitcoin price: ${formatCurrency(bitcoin)}\nEthereum price: ${formatCurrency(
+        eth
+      )}\nDogecoin price: ${formatCurrency(doge)}`
     );
   }
 };
